// src/lib/api/generated/model/documentStatus.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Myelin Base RAG API
 * API documentation for Myelin Base RAG backend
 * OpenAPI spec version: 1.0.0
 */

export type DocumentStatus =
  (typeof DocumentStatus)[keyof typeof DocumentStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const DocumentStatus = {
  PENDING_UPLOAD: "PENDING_UPLOAD",
  PROCESSING: "PROCESSING",
  COMPLETED: "COMPLETED",
  FAILED: "FAILED",
  DELETED: "DELETED",
  DELETE_FAILED: "DELETE_FAILED",
} as const;